requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
Deprecated==1.2.18
Django==4.2.19
djangorestframework==3.15.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
pycparser==2.22
PyGithub==2.5.0
PyJWT==2.10.1
PyNaCl==1.5.0
requests==2.32.3
smmap==5.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
